+++ firmware/motors.h
@@ -51,7 +51,7 @@ void set_x(uint8_t byte);
 void set_y(uint8_t byte);
 void init_motors(void);
 uint8_t motor_step(uint8_t motor, int8_t direction);
-uint8_t move_plate(int16_t dx, int16_t dy);
+uint8_t move_plate(int32_t dx, int32_t dy);
